[app] Update example for camera category.
