--- app/src/main/res/menu/menu_bounds.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/menu_action_iceland_bounds"
-        android:title="@string/restrict_iceland"
-        app:showAsAction="never"/>
-    <item
-        android:id="@+id/menu_action_allmost_world_bounds"
-        android:title="@string/restrict_almost_worldview"
-        app:showAsAction="never"/>
-    <item
-        android:id="@+id/menu_action_cross_idl"
-        android:title="@string/restrict_across_idl"
-        app:showAsAction="never"/>
-    <item
-        android:id="@+id/menu_action_reset"
-        android:title="@string/restrict_reset"
-        app:showAsAction="never"/>
-</menu>
