+++ app/src/main/res/values/example_titles.xml
@@ -27,7 +27,7 @@
     <string name="activity_location_modes">LocationModes</string>
     <string name="activity_location_tracking">Location Tracking</string>
     <string name="activity_basic_location_pulsing_circle">Basic Pulsing Circle</string>
-    <string name="activity_restrict_bounds">Restrict bounds</string>
+    <string name="activity_restrict_bounds">Restrict map panning</string>
     <string name="activity_texture_view">TextureView</string>
     <string name="activity_texture_view_animation">Animated TextureView</string>
     <string name="activity_style_circles_categorically">Style circles categorically</string>
@@ -37,7 +37,7 @@
     <string name="activity_dds_polygon_holes_title">Draw a polygon with holes</string>
     <string name="activity_dds_polygon_title">Draw a polygon</string>
     <string name="activity_dds_geojson_line_title">Draw a GeoJson line</string>
-    <string name="activity_camera_animate_title">Animate the map camera</string>
+    <string name="activity_camera_animate_title">Fly-to camera animation</string>
     <string name="activity_animated_image_source">Animated ImageSource</string>
     <string name="activity_camera_predefined_animators_title">Pre-defined animators</string>
     <string name="activity_camera_low_level">Camera Animator API</string>
